Change the module structure to match the file path with the module structure
* This is a change to support Zeitwerk
* I tried to change Concerns::IssueTemplate::Common to IssueTemplate::Common, but I changed it to IssueTemplateCommon because the IssueTemplate class existed.
