+++ app/controllers/concerns/issue_templates_common.rb
@@ -1,42 +1,40 @@
-module Concerns
-  module IssueTemplatesCommon
-    extend ActiveSupport::Concern
-    included do
-      before_action :log_action, only: [:destroy]
+module IssueTemplatesCommon
+  extend ActiveSupport::Concern
+  included do
+    before_action :log_action, only: [:destroy]
 
-      # logging action
-      def log_action
-        logger.info "[#{self.class}] #{action_name} called by #{User.current.name}" if logger
-      end
+    # logging action
+    def log_action
+      logger.info "[#{self.class}] #{action_name} called by #{User.current.name}" if logger
     end
+  end
 
-    def orphaned_templates
-      render partial: 'common/orphaned', locals: { orphaned_templates: orphaned }
-    end
+  def orphaned_templates
+    render partial: 'common/orphaned', locals: { orphaned_templates: orphaned }
+  end
 
-    def plugin_setting
-      @plugin_setting ||= Setting.plugin_redmine_issue_templates
-    end
+  def plugin_setting
+    @plugin_setting ||= Setting.plugin_redmine_issue_templates
+  end
 
-    def apply_all_projects?
-      plugin_setting['apply_global_template_to_all_projects'].to_s == 'true'
-    end
+  def apply_all_projects?
+    plugin_setting['apply_global_template_to_all_projects'].to_s == 'true'
+  end
 
-    def checklists
-      template_params[:checklists].blank? ? [] : template_params[:checklists]
-    end
+  def checklists
+    template_params[:checklists].blank? ? [] : template_params[:checklists]
+  end
 
-    def checklist_enabled?
-      Redmine::Plugin.registered_plugins.key? :redmine_checklists
-    rescue StandardError
-      false
-    end
+  def checklist_enabled?
+    Redmine::Plugin.registered_plugins.key? :redmine_checklists
+  rescue StandardError
+    false
+  end
 
-    def valid_params
-      # convert attribute name and data for checklist plugin supporting
-      attributes = template_params.except(:checklists)
-      attributes[:checklist_json] = checklists.to_json if checklist_enabled?
-      attributes
-    end
+  def valid_params
+    # convert attribute name and data for checklist plugin supporting
+    attributes = template_params.except(:checklists)
+    attributes[:checklist_json] = checklists.to_json if checklist_enabled?
+    attributes
   end
 end
