+++ app/controllers/issue_templates_controller.rb
@@ -3,8 +3,8 @@
 class IssueTemplatesController < ApplicationController
   layout 'base'
   helper :issues
-  include Concerns::IssueTemplatesCommon
-  include Concerns::ProjectTemplatesCommon
+  include IssueTemplatesCommon
+  include ProjectTemplatesCommon
   menu_item :issues
   before_action :find_tracker, :find_templates, only: %i[set_pulldown list_templates]
