+++ app/models/issue_template.rb
@@ -1,6 +1,6 @@
 class IssueTemplate < ActiveRecord::Base
   include Redmine::SafeAttributes
-  include Concerns::IssueTemplate::Common
+  include IssueTemplateCommon
   belongs_to :project
   validates :project_id, presence: true
   validates_uniqueness_of :title, scope: :project_id
